Update to store-0.4.1
This has the side effect of invalidating all of the stored binary
caches, with the upside of having better error reporting for the future.
This invalidation is also part of the fix for #3035, because the mini
build plan caches need to be invalidated since the logic for generating
them has changed.